refacto(api): split api.py in several files
